pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Danil Emurashin <emurashin@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
pydantic = "^2.10.4"
asyncpg = "^0.30.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.16"}
uvicorn = "^0.34.0"
pydantic-settings = "^2.7.1"
fastapi-async-sqlalchemy = "^0.6.1"
loguru = "^0.7.3"
greenlet = "^3.1.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi-jwt-auth = "^0.5.0"
injector = "^0.22.0"
yoyo-migrations = "^9.0.0"
fastapi-mail = "^1.4.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.6"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
